requirements.txt
google-adk==1.2.1
google-genai>=1.59.0
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic>=2.11.1,<3.0.0
google-cloud-aiplatform==1.143.0
